tests/integration/AutoMLOps/pipelines/requirements.txt
kfp<2.0.0
google-cloud-aiplatform
google-cloud-pipeline-components
google-cloud-storage
pyyaml